refactor: change 'Developer' to 'GroupMember' in file and package names
